tout bon sauf %p 9

+++ ft_printf.h
@@ -7,11 +7,12 @@
 # include <stdio.h>
 # include <string.h>
 # include <stdarg.h>
+# include <limits.h>
 
 
 int			ft_putnbr_counting(int n);
 int			ft_putchar_counting(char s);
-int			ft_putnbr_base(unsigned int n, char *s);
+int			ft_putnbr_base(unsigned long long n, char *s);
 int			ft_putstr_counting(char *s);
 int			ft_printf(const char	*fmt, ...);
 unsigned int			ft_strlen(const char	*s);
